+++ js/tests/nft.test.ts
@@ -12,14 +12,14 @@ const items = [
     owner: new PublicKey("Fxuoy3gFjfJALhwkRcuKjRdechcgffUApeYAfMWck6w8"),
     domains: [
       {
+        domain: "wallet-guide-5",
         key: "iSNVgWfb31aTWa58UxZ6fp7n3TTrUk5Gojggub5stXk",
         mint: "2RJhBbxTiPT2bZq5bhjaTZbsnhbDB7VtTAMmCdBrwBZP",
-        reverse: "wallet-guide-5",
       },
       {
+        domain: "wallet-guide-0",
         key: "uDTBDfKrJSBTgmWUZLcENPk5YrHfWbcrUbNFLjsvNpn",
         mint: "Eskv5Ns4gyREvNPPgANojNPsz6x1cbn9YwT7esAnxPhP",
-        reverse: "wallet-guide-0",
       },
     ],
   },
@@ -33,12 +33,12 @@ test("Get tokenized domains", async () => {
     )
   ).map((e) => {
     return {
+      domain: e.domain,
       key: e.key.toBase58(),
       mint: e.mint.toBase58(),
-      reverse: e.reverse,
     };
   });
-  domains.sort((a, b) => b.reverse!.localeCompare(a.reverse!));
+  domains.sort((a, b) => b.domain!.localeCompare(a.domain!));
   for (let item of items) {
     expect(domains).toEqual(item.domains);
   }